+++ HW2/solution/src/loop_pip_compiler.cpp
@@ -189,7 +189,7 @@ void LoopPipCompiler::append_regular_bundle(uint64_t instr_id, uint64_t lowest_t
     }
     
     // Place instruction in the appropriate slot based on its type
-    uint64_t bundle_idx = std::max(m_bundles.size() - 1, lowest_time);
+    uint64_t bundle_idx = std::max(static_cast<uint64_t>(m_bundles.size() - 1), lowest_time);
     
     // If we need to create a new bundle at a specific position
     while (m_bundles.size() <= bundle_idx) {
